# app/docs/swagger.yaml
definitions:
...
  response.ListData:
    properties:
      rows: {}
      total:
        type: integer
    type: object
  response.ResCode:
    enum:
    - 200
    - 401
    - 403
    - 412
    - 500
    - 600
    type: integer
    x-enum-comments:
      Error: 系统异常
      Forbidden: 没有权限
      Parameter: 参数错误
      Success: 成功
      Unauthorized: token失效
      Waring: 详情看msg
    x-enum-varnames:
    - Success
    - Unauthorized
    - Forbidden
    - Parameter
    - Error
    - Waring
  response.ResponseData:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/response.ResCode'
        description: 相应状态码
      data:
        description: 数据
      msg:
        description: 提示信息
        type: string
    type: object
  systemModels.Accredit:
    properties:
      posts:
        description: 岗位
        items:
          $ref: '#/definitions/systemModels.SysPostVo'
        type: array
      roles:
        description: 角色
        items:
          $ref: '#/definitions/systemModels.SysRoleVo'
        type: array
    type: object
...
  systemModels.EditUserStatus:
    properties:
      createBy:
        description: 创建人
        type: integer
      createTime:
        description: 创建时间
        type: string
      status:
        description: 状态
        type: string
      updateBy:
        description: 修改人
        type: integer
      updateTime:
        description: 修改时间
        type: string
      userId:
        description: 用户id
        example: "0"
        type: string
    required:
    - status
    - userId
    type: object
...
  systemModels.ResetPwd:
    properties:
      password:
        description: 新密码
        type: string
      userId:
        description: 用户ID
        example: "0"
        type: string
    required:
    - password
    - userId
    type: object
...
  systemModels.SysPostVo:
    properties:
      createBy:
        description: 创建人
        type: integer
      createTime:
        description: 创建时间
        type: string
      postCode:
        type: string
      postId:
        example: "0"
        type: string
      postName:
        type: string
      postSort:
        type: integer
      remark:
        type: string
      status:
        type: string
      updateBy:
        description: 修改人
        type: integer
      updateTime:
        description: 修改时间
        type: string
    type: object
...
  systemModels.SysRoleVo:
    properties:
      createBy:
        description: 创建人
        type: integer
      createTime:
        description: 创建时间
        type: string
      delFlag:
        type: string
      remark:
        type: string
      roleId:
        example: "0"
        type: string
      roleKey:
        type: string
      roleName:
        type: string
      roleSort:
        type: integer
      status:
        type: string
      updateBy:
        description: 修改人
        type: integer
      updateTime:
        description: 修改时间
        type: string
    type: object
  systemModels.SysUserDML:
    properties:
      avatar:
        description: 头像
        type: string
      createBy:
        description: 创建人
        type: integer
      createTime:
        description: 创建时间
        type: string
      dataScope:
        description: 权限范围
        type: string
      deptId:
        description: 部门ID
        example: "0"
        type: string
      email:
        description: 邮箱
        type: string
      nickName:
        description: 用户昵称
        type: string
      password:
        description: 密码
        type: string
      phonenumber:
        description: 手机号
        type: string
      postIds:
        description: 岗位IDS
        items:
          type: string
        type: array
      remark:
        description: 备注
        type: string
      roleIds:
        description: 角色IDS
        items:
          type: string
        type: array
      sex:
        description: 性别
        type: string
      status:
        description: 状态
        type: string
      updateBy:
        description: 修改人
        type: integer
      updateTime:
        description: 修改时间
        type: string
      userName:
        description: 用户名
        type: string
    required:
    - deptId
    - nickName
    - password
    - sex
    - userName
    type: object
  systemModels.SysUserDataScope:
    properties:
      dataScope:
        description: 数据范围（1：全部数据权限 2：自定数据权限 3：本部门数据权限 4：本部门及以下数据权限,无任何）权限
        type: string
      deptIds:
        description: 如果是自定义就是部门ID 其他不填
        items:
          type: string
        type: array
      userId:
        description: 用户ID
        example: "0"
        type: string
    required:
    - dataScope
    - userId
    type: object
...
  systemModels.SysUserVo:
    properties:
      avatar:
        type: string
      createBy:
        description: 创建人
        type: integer
      createTime:
        description: 创建时间
        type: string
      dataScope:
        type: string
      delFlag:
        type: string
      deptId:
        example: "0"
        type: string
      deptName:
        type: string
      email:
        type: string
      leader:
        type: string
      nickName:
        type: string
      phonenumber:
        type: string
      remark:
        type: string
      roleId:
        type: integer
      sex:
        type: string
      status:
        type: string
      updateBy:
        description: 修改人
        type: integer
      updateTime:
        description: 修改时间
        type: string
      userId:
        example: "0"
        type: string
      userName:
        type: string
    type: object
...
  systemModels.UserAndAccredit:
    properties:
      postIds:
        description: 选择的岗位Id
        items:
          type: string
        type: array
      posts:
        description: 岗位
        items:
          $ref: '#/definitions/systemModels.SysPostVo'
        type: array
      roleIds:
        description: 选择的角色Id
        items:
          type: string
        type: array
      roles:
        description: 角色
        items:
          $ref: '#/definitions/systemModels.SysRoleVo'
        type: array
      user:
        allOf:
        - $ref: '#/definitions/systemModels.SysUserVo'
        description: user
    type: object
...
info:
  contact:
    email: zhao_402295440@126.com
    name: danny
  description: baize接口文档
  title: baize
  version: 2.0.x
paths:
...
  /system/user:
    get:
      description: 查询用户列表
      parameters:
      - description: 注册开始时间
        in: query
        name: beginTime
        type: string
      - description: 部门ID
        in: query
        name: deptId
        type: integer
      - description: 注册结束时间
        in: query
        name: endTime
        type: string
      - description: 排序规则  降序desc   asc升序
        in: query
        name: isAsc
        type: string
      - description: 排序字段
        in: query
        name: orderBy
        type: string
      - default: 1
        description: 第几页
        in: query
        name: pageNum
        type: integer
      - default: 10000
        description: 数量
        in: query
        name: pageSize
        type: integer
      - description: 电话
        in: query
        name: phonenumber
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - description: 用户名
        in: query
        name: userName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.ListData'
                  - properties:
                      Rows:
                        items:
                          $ref: '#/definitions/systemModels.SysUserVo'
                        type: array
                    type: object
              type: object
      security:
      - BearerAuth: []
      summary: 查询用户列表
      tags:
      - 用户相关
    post:
      description: 添加用户
      parameters:
      - description: 用户信息
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/systemModels.SysUserDML'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 添加用户
      tags:
      - 用户相关
    put:
      description: 修改用户
      parameters:
      - description: 用户信息
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/systemModels.SysUserDML'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 修改用户
      tags:
      - 用户相关
  /system/user/:
    get:
      description: 获取当前用户信息
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/systemModels.Accredit'
              type: object
      security:
      - BearerAuth: []
      summary: 获取当前用户信息
      tags:
      - 用户相关
  /system/user/:userIds:
    delete:
      description: 删除用户
      parameters:
      - collectionFormat: csv
        description: userIds
        in: path
        items:
          type: integer
        name: userIds
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 删除用户
      tags:
      - 系统用户
  /system/user/{userId}:
    get:
      description: 根据用户ID获取用户信息
      parameters:
      - description: userId
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/systemModels.UserAndAccredit'
              type: object
      security:
      - BearerAuth: []
      summary: 根据用户ID获取用户信息
      tags:
      - 用户相关
  /system/user/authRole:
    put:
      description: 授权角色
      parameters:
      - description: 角色id
        in: query
        name: string
        required: true
        type: string
      - description: 用户id
        in: query
        name: string
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.ListData'
                  - properties:
                      Rows:
                        items:
                          $ref: '#/definitions/systemModels.SysUserVo'
                        type: array
                    type: object
              type: object
      security:
      - BearerAuth: []
      summary: 授权角色
      tags:
      - 用户相关
  /system/user/authRole/{userId}:
    get:
      description: 根据用户编号获取授权角色
      parameters:
      - description: userId
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: 根据用户编号获取授权角色
      tags:
      - 用户相关
  /system/user/changeStatus:
    put:
      description: 修改用户状态
      parameters:
      - description: 用户信息
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/systemModels.EditUserStatus'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 修改用户状态
      tags:
      - 用户相关
  /system/user/dataScope:
    put:
      description: 修改数据权限
      parameters:
      - description: 用户权限信息
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/systemModels.SysUserDataScope'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 修改数据权限
      tags:
      - 用户相关
  /system/user/dataScope/{userId}:
    get:
      description: 查询用户数据权限
      parameters:
      - description: userId
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 查询用户数据权限
      tags:
      - 用户相关
  /system/user/export:
    post:
      description: 导出用户
      parameters:
      - description: 注册开始时间
        in: query
        name: beginTime
        type: string
      - description: 部门ID
        in: query
        name: deptId
        type: integer
      - description: 注册结束时间
        in: query
        name: endTime
        type: string
      - description: 排序规则  降序desc   asc升序
        in: query
        name: isAsc
        type: string
      - description: 排序字段
        in: query
        name: orderBy
        type: string
      - default: 1
        description: 第几页
        in: query
        name: pageNum
        type: integer
      - default: 10000
        description: 数量
        in: query
        name: pageSize
        type: integer
      - description: 电话
        in: query
        name: phonenumber
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - description: 用户名
        in: query
        name: userName
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            items:
              type: integer
            type: array
      security:
      - BearerAuth: []
      summary: 导出用户
      tags:
      - 系统用户
  /system/user/importData:
    post:
      consumes:
      - multipart/form-data
      description: 导入用户
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 导入用户
      tags:
      - 系统用户
  /system/user/input:
    post:
      description: 导出用户
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: integer
            type: array
      security:
      - BearerAuth: []
      summary: 导出用户
      tags:
      - 系统用户
...
  /system/user/resetPwd:
    put:
      description: 重置密码
      parameters:
      - description: 密码
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/systemModels.ResetPwd'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 重置密码
      tags:
      - 用户相关
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"